Refactored Lexer into Lexer and separate Tokenizer
Introduced a `Tokenizer` class that is used to keep track of the state
during the lexing of the code. It is the functional core of the
application, all objects are immutable and frozen when after
initialization.

Added specific tokenizer classes for converting the chunks of code as
matched by the `Lexer`. Naming them tokenizers might not be correct,
have to think about that.